Revert "introduced PSLP_uint"
This reverts commit 9dafbef89cba7c1fe004e00c6005b9d6e29b1aff.

+++ include/PSLP/PSLP_sol.h
@@ -31,8 +31,8 @@ extern "C"
         double *x;
         double *y;
         double *z;
-        int dim_x;
-        int dim_y;
+        size_t dim_x;
+        size_t dim_y;
     } Solution;
 
 #ifdef __cplusplus
